Added JSON as alternative storage. Ajax bookmark save. Added error box. Few UX updates.

+++ index.php
@@ -5,7 +5,9 @@ $current_dir_url = "http://" . $_SERVER['HTTP_HOST'] . dirname($_SERVER['SCRIPT_
 list( $directories, $files ) = $pi->getDirContents();
 $local_addr = $pi->getLocalAddress();
 $public_addr = true ? $pi->getPublicAddress() : 'N/A';
+$isBookmarksEnabled = !empty($pi->storage);
 $bookmarks = $pi->getBookmarks();
+$errors = $pi->errors;
 
 ?>
 
@@ -27,6 +29,8 @@ $bookmarks = $pi->getBookmarks();
       a { text-decoration: none !important; background: none !important; }
       a:hover { color: red; }
       p { margin: 0; padding: 2px 0; }
+      svg { padding-top: 2px; }
+      .errorbox p { color: red; }
       .container { font-family: monospace; max-width: 768px; background: #DDD; padding: 10px; margin: auto; }
       .center { text-align: center; }
       .text-left { text-align: left; }
@@ -43,12 +47,11 @@ $bookmarks = $pi->getBookmarks();
       .current-dir:hover { background-color: lightblue; }
       .bm-list-wrapper { overflow: hidden; }
       ul { padding: 0; margin: 0; }
-      .bookmarks { display: inline-block; padding: 0 5px 5px 5px; }
+      .bookmarks { display: inline-block; padding: 1px 4px 2px 4px;}
       .exterminate a { color: red; }
       .exterminate .bookmarks:hover { text-decoration: line-through; }
       #bm-form-wrapper { padding-top: 2px; padding-bottom: 4px; }
       #bm-list{ max-width: 700px; margin-left: 10px; }
-      #bookmark_form{ text-align: center; }
       ::-webkit-scrollbar { width: 6px; height: 6px; } /* Track */
       ::-webkit-scrollbar-track { -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3); -webkit-border-radius: 10px; border-radius: 10px; }
        /* Handle */
@@ -71,18 +74,21 @@ $bookmarks = $pi->getBookmarks();
       window.onload = function() {
         function bookmark_delete(){
           var bm_list = document.getElementById('bookmarks-list');
+          if( bm_list == null ) return;
           bm_list.addEventListener('click', function(event) {
             if (bm_list.className == "exterminate") {
               var clickedEl = event.target;
               if(clickedEl.tagName === 'A') {
                 event.preventDefault();
-                var listItem = clickedEl.parentNode;
-                listItem.parentNode.removeChild(listItem);
-                var bm_id = clickedEl.getAttribute('data-bookmark-id');
-                var xhttp = new XMLHttpRequest();
-                xhttp.open("POST", window.location.href, true);
-                xhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
-                xhttp.send("delete=" + bm_id);
+                if( confirm('Delete bookmark "' + clickedEl.innerHTML + '"?') ) {
+                  var listItem = clickedEl.parentNode;
+                  listItem.parentNode.removeChild(listItem);
+                  var bm_id = clickedEl.getAttribute('data-bookmark-id');
+                  var xhttp = new XMLHttpRequest();
+                  xhttp.open("POST", window.location.href, true);
+                  xhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+                  xhttp.send("delete=" + bm_id);
+                }
               }
             }
           });
@@ -178,9 +184,9 @@ $bookmarks = $pi->getBookmarks();
         document.getElementById("eval_output").style.display='none';
         document.getElementById("info_wrapper").style.display='block';
       }
-      function collapseDirContents() {
-        var wrapper = document.getElementById('directory_index');
-        wrapper.style.display = wrapper.style.display === 'none' ? 'block' : 'none';
+      function toggleElementView(elementID) {
+        var elem = document.getElementById(elementID);
+        elem.style.display = elem.style.display === 'none' ? 'block' : 'none';
       }
       function webSearch(engine) {
         var search_form = document.search_form;
@@ -199,18 +205,34 @@ $bookmarks = $pi->getBookmarks();
               search_form.action = "//www.google.com/search";
         }
       }
-      function show_bm_block(){
-        var bm_form_wrapper = document.getElementById('bm-form-wrapper');
-        bm_form_wrapper.style.display = bm_form_wrapper.style.display === 'none' ? 'block' : 'none';
-      }
       function remove_btn(){
         var bm_list = document.getElementById('bookmarks-list');
         bm_list.classList.toggle('exterminate');
+      }
+      function bookmarkSave() {
+        var title = document.getElementById('bookmark_name').value;
+        var url = document.getElementById('bookmark_url').value;
+        var xhttp = new XMLHttpRequest();
+        xhttp.onreadystatechange = function() {
+          if (this.readyState == 4 && this.status == 200) {
+            location.reload();
+          }
+        };
+        xhttp.open("POST", window.location.href, true);
+        xhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+        xhttp.send( "addbookmark=1&bookmark_name=" + title + "&bookmark_url=" + encodeURI(url) );
       }
     </script>
   </head>
   <body>
     <div class="container">
+      <?php if( !empty($errors) ) { ?>
+        <div class="block center errorbox">
+          <?php foreach( $errors as $error ) { ?>
+            <p><?php print $error; ?></p>
+          <?php } ?>
+        </div>
+      <?php } ?>
       <div class="block center">
         <a href="javascript:location.reload();"><h1>localhost | <?php echo getHostByName(getHostName()); ?></h1></a>
       </div>
@@ -246,31 +268,35 @@ $bookmarks = $pi->getBookmarks();
             </form>
           </div>
         </div>
-        <div id="bm-form-wrapper" class="block" style="display: none;">
-          <form id="bookmark_form" method="post" action="">
-            <label for="bookmark_name">Title: </label><input type="text" name="bookmark_name" id="bookmark_name" placeholder="Enter bookmark title">
-            <label for="bookmark_url">URL: </label><input type="text" name="bookmark_url" id="bookmark_url" placeholder="Enter URL to bookmark">
-            <input type="submit" name="submit" value="Bookmark">
-          </form>
+        <?php if($isBookmarksEnabled) { ?>
+        <div id="bm-form-wrapper" class="block center" style="display: none;">
+          <label for="bookmark_name">Title: </label><input type="text" name="bookmark_name" id="bookmark_name" placeholder="Enter bookmark title">
+          <label for="bookmark_url">URL: </label><input type="text" name="bookmark_url" id="bookmark_url" placeholder="Enter URL to bookmark">
+          <input type="button" id="bookmark_save" name="bookmark_save" onclick="bookmarkSave();" value="Save Bookmark">
         </div>
         <div class="block" id="bookmark_wrapper">
           <div class="bm-list-wrapper">
             <div class="left">
-              <a href="javascript:;" onclick="show_bm_block()"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="16" height="16" viewBox="0 0 32 32"><path d="M31 12h-11v-11c0-0.552-0.448-1-1-1h-6c-0.552 0-1 0.448-1 1v11h-11c-0.552 0-1 0.448-1 1v6c0 0.552 0.448 1 1 1h11v11c0 0.552 0.448 1 1 1h6c0.552 0 1-0.448 1-1v-11h11c0.552 0 1-0.448 1-1v-6c0-0.552-0.448-1-1-1z"></path></svg></a>
+              <a href="javascript:;" onclick="toggleElementView('bm-form-wrapper')" title="Add Bookmark"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="13" height="13" viewBox="0 0 32 32"><path d="M31 12h-11v-11c0-0.552-0.448-1-1-1h-6c-0.552 0-1 0.448-1 1v11h-11c-0.552 0-1 0.448-1 1v6c0 0.552 0.448 1 1 1h11v11c0 0.552 0.448 1 1 1h6c0.552 0 1-0.448 1-1v-11h11c0.552 0 1-0.448 1-1v-6c0-0.552-0.448-1-1-1z"></path></svg></a>
             </div>
             <div id="bm-list" style="display: inline-block;">
               <ul id="bookmarks-list">
-                <?php foreach($bookmarks as $row) { ?>
-                  <li class="bookmarks"><a href="<?php echo $row['url'];?>" data-bookmark-id="<?php echo $row['id'];?>" target="_blank"><?php echo $row['title'];?></a></li>
+                <?php if( count($bookmarks) > 0 ) { ?>
+                  <?php foreach($bookmarks as $bookmark) { ?>
+                    <li class="bookmarks"><a href="<?php echo $bookmark->url;?>" data-bookmark-id="<?php echo $bookmark->id;?>" target="_blank"><?php echo $bookmark->title;?></a></li>
+                  <?php } ?>
+                <?php } else { ?>
+                  <li class="bookmarks bookmark-help">No bookmarks. Add bookmarks using the <b>+</b> button.</li>
                 <?php } ?>
               </ul>
             </div>
             <div class="right">
-              <a href="javascript:;" onclick="remove_btn()"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="16" height="16" viewBox="0 0 32 32"><path d="M31.708 25.708c-0-0-0-0-0-0l-9.708-9.708 9.708-9.708c0-0 0-0 0-0 0.105-0.105 0.18-0.227 0.229-0.357 0.133-0.356 0.057-0.771-0.229-1.057l-4.586-4.586c-0.286-0.286-0.702-0.361-1.057-0.229-0.13 0.048-0.252 0.124-0.357 0.228 0 0-0 0-0 0l-9.708 9.708-9.708-9.708c-0-0-0-0-0-0-0.105-0.104-0.227-0.18-0.357-0.228-0.356-0.133-0.771-0.057-1.057 0.229l-4.586 4.586c-0.286 0.286-0.361 0.702-0.229 1.057 0.049 0.13 0.124 0.252 0.229 0.357 0 0 0 0 0 0l9.708 9.708-9.708 9.708c-0 0-0 0-0 0-0.104 0.105-0.18 0.227-0.229 0.357-0.133 0.355-0.057 0.771 0.229 1.057l4.586 4.586c0.286 0.286 0.702 0.361 1.057 0.229 0.13-0.049 0.252-0.124 0.357-0.229 0-0 0-0 0-0l9.708-9.708 9.708 9.708c0 0 0 0 0 0 0.105 0.105 0.227 0.18 0.357 0.229 0.356 0.133 0.771 0.057 1.057-0.229l4.586-4.586c0.286-0.286 0.362-0.702 0.229-1.057-0.049-0.13-0.124-0.252-0.229-0.357z"></path></svg></a>
+              <a href="javascript:;" onclick="remove_btn()" title="Exterminator! Click me and then click on any bookmark to delete it"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="13" height="13" viewBox="0 0 32 32"><path d="M31.708 25.708c-0-0-0-0-0-0l-9.708-9.708 9.708-9.708c0-0 0-0 0-0 0.105-0.105 0.18-0.227 0.229-0.357 0.133-0.356 0.057-0.771-0.229-1.057l-4.586-4.586c-0.286-0.286-0.702-0.361-1.057-0.229-0.13 0.048-0.252 0.124-0.357 0.228 0 0-0 0-0 0l-9.708 9.708-9.708-9.708c-0-0-0-0-0-0-0.105-0.104-0.227-0.18-0.357-0.228-0.356-0.133-0.771-0.057-1.057 0.229l-4.586 4.586c-0.286 0.286-0.361 0.702-0.229 1.057 0.049 0.13 0.124 0.252 0.229 0.357 0 0 0 0 0 0l9.708 9.708-9.708 9.708c-0 0-0 0-0 0-0.104 0.105-0.18 0.227-0.229 0.357-0.133 0.355-0.057 0.771 0.229 1.057l4.586 4.586c0.286 0.286 0.702 0.361 1.057 0.229 0.13-0.049 0.252-0.124 0.357-0.229 0-0 0-0 0-0l9.708-9.708 9.708 9.708c0 0 0 0 0 0 0.105 0.105 0.227 0.18 0.357 0.229 0.356 0.133 0.771 0.057 1.057-0.229l4.586-4.586c0.286-0.286 0.362-0.702 0.229-1.057-0.049-0.13-0.124-0.252-0.229-0.357z"></path></svg></a>
             </div>
           </div>
         </div>
-        <a onclick="collapseDirContents();" href="javascript:;">
+        <?php } ?>
+        <a onclick="toggleElementView('directory_index');" href="javascript:;">
           <div class="block clearfix center current-dir">Current directory: <?php echo __DIR__; ?></div>
         </a>
         <div id="directory_index" class="block directory_index clearfix" style="display: none;">
@@ -315,7 +341,11 @@ $bookmarks = $pi->getBookmarks();
 
 class Pi {
   const DB = 'pretty_index.sqlite';
+  const FS = 'pretty_index.json';
   private $_pdo = null;
+  private $_jsonData = null;
+  public $storage = null;
+  public $errors = array();
 
   function __construct() {
     $this->_setupStorage();
@@ -326,8 +356,9 @@ class Pi {
     if ( !empty($_REQUEST['phpinfo']) and $_REQUEST['phpinfo'] == 1 ) {
       phpinfo();
       exit;
-    } else if( !empty($_POST['bookmark_name']) and !empty($_POST['bookmark_url']) ) {
-      $this->saveBookmark( $_POST['bookmark_name'], $_POST['bookmark_url'] );
+    } else if( !empty($_REQUEST['addbookmark']) and !empty($_REQUEST['bookmark_name']) and !empty($_REQUEST['bookmark_url']) ) {
+      $this->saveBookmark( $_REQUEST['bookmark_name'], $_REQUEST['bookmark_url'] );
+      exit;
     } else if ( !empty($_REQUEST['delete']) ) {
       $this->deleteBookmark($_REQUEST['delete']);
     } else if( !empty($_POST['eval_code']) and isset( $_SERVER['HTTP_X_REQUESTED_WITH'] ) ) {
@@ -343,20 +374,41 @@ class Pi {
 
   private function _setupStorage() {
     try {
-      $this->_pdo = new PDO( "sqlite:" . self::DB );
-      $this->_pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-      $this->_pdo->exec("CREATE TABLE IF NOT EXISTS bookmarks (id INTEGER PRIMARY KEY, title TEXT NOT NULL UNIQUE, url TEXT NOT NULL UNIQUE)");
+      if( is_writable(__DIR__) and in_array('sqlite', PDO::getAvailableDrivers()) ) {
+        $this->_pdo = new PDO( "sqlite:" . self::DB );
+        $this->_pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+        $this->_pdo->exec("CREATE TABLE IF NOT EXISTS bookmarks (id INTEGER PRIMARY KEY, title TEXT NOT NULL UNIQUE, url TEXT NOT NULL UNIQUE)");
+        $this->storage = self::DB;
+      } else if( touch(self::FS) and is_writable(self::FS) ) {
+        $json = json_decode( file_get_contents(self::FS) );
+        $jsonDB = new stdClass();
+        $indices = new stdClass();
+        $indices->bookmarksIndex = 0;
+        $jsonDB->indices = $indices;
+        $jsonDB->bookmarks = array();
+        $this->_jsonData = !empty($json) ? $json : $jsonDB;
+        $this->storage = self::FS;
+      } else {
+        $this->errors[] = "Bookmark feature requires pdo-sqlite extension and write permission on " . __DIR__ . " directory.";
+        $this->errors[] = "Alternatively, a writable JSON file named " . self::FS . " will do.";
+      }
     } catch( PDOException $e ) {
-      echo $e->getMessage();
+      $this->errors[] = $e->getMessage();
     }
   } // end of _setupStorage()
 
   private function _query( $sql ) {
     try {
-      return !empty($this->_pdo) ? $this->_pdo->query($sql) : false;
+      return !empty($this->_pdo) ? $this->_pdo->query($sql, PDO::FETCH_CLASS, "stdClass") : false;
     } catch (Exception $e) {
-      echo $e->getMessage();
+      $this->errors[] = $e->getMessage();
     }
+  }
+
+  private function _writeJSON( $option = JSON_PRETTY_PRINT ) {
+    if( !empty($this->_jsonData) )
+      return @file_put_contents( self::FS, json_encode($this->_jsonData, $option) );
+    return false;
   }
 
   public function getLocalAddress() {
@@ -406,15 +458,41 @@ class Pi {
   }
 
   public function saveBookmark($title, $url) {
-    $this->_query("INSERT INTO bookmarks (id, title, url) VALUES (NULL, '$title', '$url')");
+    if( empty($this->storage) ) { return; }
+    if( $this->storage === self::DB ) {
+      $title = $this->_pdo->quote($title);
+      $url = $this->_pdo->quote($url);
+      $this->_query("INSERT INTO bookmarks (id, title, url) VALUES (NULL, $title, $url)");
+    } else {
+      $newBookmark = new stdClass();
+      $newBookmark->id = ++$this->_jsonData->indices->bookmarksIndex;
+      $newBookmark->title = $title;
+      $newBookmark->url = $url;
+      $this->_jsonData->bookmarks[] = $newBookmark;
+      $this->_writeJSON(0);
+    }
   }
 
   public function deleteBookmark($id) {
-    $this->_query("DELETE FROM bookmarks WHERE id = $id");
+    if( empty($this->storage) ) { return; }
+    if( $this->storage === self::DB ) {
+      $this->_query("DELETE FROM bookmarks WHERE id = $id");
+    } else {
+      foreach( $this->_jsonData->bookmarks as $index => $bookmark ) {
+        if( $bookmark->id == $id ) {
+          unset( $this->_jsonData->bookmarks[$index] );
+          $this->_jsonData->bookmarks = array_values($this->_jsonData->bookmarks); // important to reindex
+          $this->_writeJSON(0);
+          break;
+        }
+      } // end foreach
+    }
   }
 
   public function getBookmarks() {
-    return $this->_query("SELECT * FROM bookmarks;");
+    if( empty($this->storage) ) { return array(); }
+    $bm = ( $this->storage === self::DB ) ? $this->_query("SELECT * FROM bookmarks")->fetchAll() : $this->_jsonData->bookmarks;
+    return count($bm) > 0 ? $bm : array();
   }
 
 } // end of class Pi
